LayoutTests/vibration/navigator-vibration-expected.txt

Test for the Vibration API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.vibrate(0); is undefined
PASS navigator.vibrate([]); is undefined
PASS navigator.vibrate(1000); is undefined
PASS navigator.vibrate([1000, 300, 500]); is undefined
PASS navigator.vibrate(4294967295); is undefined
PASS navigator.vibrate(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
